pyproject.toml
[project]
name = "django-starterkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.2",
    "django-filter>=24.3",
    "django-phonenumber-field[phonenumberslite]>=8.0.0",
    "django>=5.1.1",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular[sidecar]>=0.27.2",
    "gunicorn>=23.0.0",
    "markdown>=3.7",
    "uvicorn[standard]>=0.31.0",
    "django-environ>=0.11.2",
    "psycopg[binary]>=3.2.3",
    "django-cache-url>=3.4.5",
    "dj-database-url>=2.2.0",
    "bpython>=0.24",
    "ruff>=0.6.9",
    "factory-boy>=3.3.1",
    "pytest-django>=4.9.0",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=5.0.0",
    "coverage>=7.6.1",
    "pillow>=10.4.0",
    "whitenoise>=6.7.0",
    "redis>=5.1.0",
    "celery>=5.4.0",
    "flower>=2.0.1",
    "django-celery-beat>=2.7.0",
    "django-redis>=5.4.0",
    "django-celery-results>=2.5.1",
    "django-storages[s3]>=1.14.4",
    "python-json-logger>=2.0.7",
    "django-silk>=5.2.0",
    "django-countries>=7.6.1",
]

[tool.uv]
dev-dependencies = ["pyclean>=3.0.0"]

[tool.coverage.run]
branch = true
omit = [
    "*/urls/*",
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/tests.py",
    "*/apps.py",
    "*/rss.py",
    "*/urls.py",
    "manage.py",
    "settings.py",
    "asgi.py",
    "wsgi.py",
    "gunicorn_config.py",
    "storages.py"
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "def __unicode__",
    "if DEBUG",
    "if TYPE_CHECKING",
    "def __repr__",
    "def __str__",
]

[tool.pytest]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = "tests.py test_*.py *_tests.py"
addopts = "-v --nomigrations --ignore=venv --cov=. --cov-report=html"

[tool.uv.sources]
django-starterkit = { workspace = true }
